Service App Added (Dynamic Content)
Added dynamic content which will be shown through service.html template. In the admin page, an admin can create instances of the model Service, and said model will be displayed in the service page dynamically. In other word, when we create or delete new Services in the admin page, these changes will be reflected in the public website. The model has serveral fields, for instance, an ImageField (this requires a python package called Pillow).